ProgressBarの表示を切り替えるコードをクラスに切り出す
